+++ games_hub/static.py
@@ -5,7 +5,6 @@ PROJECT_NAME = "GamesHub"
 GITHUB_URL = "https://github.com/lupohan44/GamesHub"
 GITHUB_VERSION_URL = "https://raw.githubusercontent.com/lupohan44/GamesHub/main/version.py"
 GITHUB_MIRROR_VERSION_URL = "https://ghproxy.com/" + GITHUB_VERSION_URL
-STEAM_DB_FREE_GAMES_URL = "https://steamdb.info/upcoming/free/"
 CONFIG_PATH = "config.json5"
 
 # log format